Refactor game modes into separate modules
