+++ src/components/word-association/modes/phrasalVerb.ts
@@ -1,0 +1,4 @@
+import type { WordGroup, WordItem, GameRound } from '../types';
+
+export { fetchIdiomGroups as fetchPhrasalVerbGroups } from './idiom';
+export { buildIdiomRound as buildPhrasalVerbRound } from './idiom';
